requirements.txt
numpy>=1.21.0
matplotlib>=3.4.3
h5py>=3.2.1
pillow>=8.2.0
scipy>=1.7.0
